SHA256: add range-checks to make gadget sound
The [aligned_variables] from [split_and_rotate], used
in the SHA256 gadget, have to be u4s.
For most cases, this was done by the call to xor.
But for the ones stored in [t1_shifted_10], some  variables
are not checked, as they are overwritten.

This commit adds the missing checks.